Improve how we handle profiles across multiple things
