+++ Cargo.toml
@@ -16,6 +16,7 @@ members = [
     "kanidm_unix_int/nss_kanidm",
     "kanidm_unix_int/pam_kanidm",
     "orca",
+    "profiles",
 ]
 
 exclude = [
