+++ kanidm_unix_int/Cargo.toml
@@ -84,7 +84,5 @@ score = { path = "../kanidmd/score" }
 
 [build-dependencies]
 structopt = { version = "0.3", default-features = false }
-toml = "0.5"
-serde = "1.0"
-serde_derive = "1.0"
+profiles = { path = "../profiles" }
