+++ kanidmd/daemon/Cargo.toml
@@ -25,7 +25,6 @@ tokio = { version = "1", features = ["full"] }
 toml = "0.5"
 
 [build-dependencies]
+serde = { version = "1", features = ["derive"] }
 structopt = { version = "0.3", default-features = false }
-serde = "1.0"
-serde_derive = "1.0"
-toml = "0.5"
+profiles = { path = "../../profiles" }
