+++ profiles/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "profiles"
+version = "1.1.0"
+authors = ["William Brown <william@blackhats.net.au>"]
+edition = "2018"
+license = "MPL-2.0"
+description = "Kanidm Build System Profiles"
+documentation = "https://docs.rs/kanidm/latest/kanidm/"
+homepage = "https://github.com/kanidm/kanidm/"
+repository = "https://github.com/kanidm/kanidm/"
+
+[lib]
+name = "profiles"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+toml = "0.5"
+rustc_version = "0.4.0"
